removed some ugliness
git-svn-id: file:///tmp/owch2_2/branches/hyperarchy@9 afdb5d6e-910f-0410-ac8a-c8b07d4428ab

+++ java/net/sourceforge/owch2/kernel/ListenerFactory.java
@@ -3,7 +3,6 @@ package net.sourceforge.owch2.kernel;
 import java.net.*;
 import java.util.concurrent.*;
 
-
 /*
 *
 * ListenerFactory
@@ -58,7 +57,7 @@ abstract public class ListenerFactory {
     }
 
     public void setThreads(Integer threads) {
-        this.threads = threads;
+        this.threads = threads.intValue();
     }
 
     public void setSocketCount(int socketCount) {
